Move all countries methods to a separate resource
