+++ CakeMail.RestClient.IntegrationTests/Countries.cs
@@ -11,13 +11,13 @@ namespace CakeMail.RestClient.IntegrationTests
 			Console.WriteLine(new string('-', 25));
 			Console.WriteLine("Executing COUNTRIES methods...");
 
-			var countries = api.GetCountries();
+			var countries = api.Countries.GetList();
 			Console.WriteLine("Retrieved all countries. There are {0} countries.", countries.Count());
 
 			var canada = countries.Single(country => country.EnglishName == "Canada");
 			Console.WriteLine("Canada --> Id: {0}", canada.Id);
 
-			var canadianProvinces = api.GetProvinces(canada.Id);
+			var canadianProvinces = api.Countries.GetProvinces(canada.Id);
 			Console.WriteLine("There are {0} canadian provinces/territories/etc.", canadianProvinces.Count());
 
 			var quebec = canadianProvinces.Single(province => province.EnglishName == "Quebec");
@@ -26,7 +26,7 @@ namespace CakeMail.RestClient.IntegrationTests
 			var usa = countries.Single(country => country.EnglishName == "United States");
 			Console.WriteLine("USA --> Id: {0}", usa.Id);
 
-			var americanStates = api.GetProvinces(usa.Id);
+			var americanStates = api.Countries.GetProvinces(usa.Id);
 			Console.WriteLine("There are {0} american states/territories/etc.", americanStates.Count());
 
 			var georgia = americanStates.Single(province => province.EnglishName == "Georgia");
