+++ CakeMail.RestClient/CakeMail.RestClient.csproj
@@ -106,6 +106,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\Campaigns.cs" />
     <Compile Include="Resources\Clients.cs" />
+    <Compile Include="Resources\Countries.cs" />
     <Compile Include="Utilities\CakeMailIntegerBooleanConverter.cs" />
     <Compile Include="Utilities\Constants.cs" />
     <Compile Include="Utilities\ExtensionMethods.cs" />
